remove new task fields and undroppable week
